requirements.txt
aqtinstall==2.0.0
flask
flask_cors
gunicorn
